internal/querytext: new package that extracts the parameter parser
Prior to v1, we should remove the query pre-parser from the main driver.
Also prior to v1, we can create a new Connector that does the processing.
